Added sequence scatter plot, REQUIRES BIOINFORMATICS TOOLBOX
Added a sequence scatter plot by using MDS dimensional reduction on a computed Jukes-Cantor distance matrix. All sequences that were detected are used. Initial conditions are changed so that the two initial (very different) sequences have an initially high viral count. This is done in order to show clear clustering in the scatter plot. The calculation of the Jukes-Cantor distance matrix require the bioinformatics toolbox, which first needs to be installed to plot the scatter plot.