Added immune selection
Immune cells are created, can kill infected cells and decay. This is done by replacing reaction 3 with 3 reactions, as mentioned in Woo and Reifman 2014 (also 2012). Immune cells are specific for a genotype (nucleotide sequence), eventually this should be changed so that immune cells are specific to phenotype (amino acid sequence). I also made some other changes (e.g. to the remove function).