Added immune reactions
Added the 3 immune reactions of T-cell stimulation by infected cells, T-cell killing of infected cells and T-cell clearance.
Firstly, this implementation still lacks separate tracking of immune cells, instead immune cells are removed when all strains to which that immune cells is specific dissapear. This means that if the epitope reemerges shortly after it will not be immediately recognized. This might be a rare situation and OK to ignore.
Secondly, this implementation lacks 'real epitopes' and instead defines the entire AA sequence as an epitope.
If stim=0 then no T-cells are created and the normal simulation should be done.